requirements.txt
py_clob_client
python-dotenv
websockets
web3
